Вынос функции таймера в отдельный модуль
